Update reworked on ls,rm,listdir,mvb command

+++ src/commands/Command_List.java
@@ -13,6 +13,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Stream;
 
 public class Command_List {
@@ -55,21 +56,33 @@ public class Command_List {
     public static List<String> listDirectory() {
         List<String> dic = new ArrayList<>();
         Path path = Paths.get(Os_res.system_path+pwdCommand());
-        try (Stream<Path> fileList =Files.walk(path)){
-            fileList.filter(Files::isDirectory).forEach(file->{
-                dic.add(file.getFileName().toString());
-            });
-        }
-        catch (IOException e ){
+        try(Stream<Path> list=Files.list(path)){
+            list.filter(Files::isDirectory)
+                    .map(Path::getFileName).forEach(
+                            fileName -> {
+                                dic.add(fileName.toString());
+                            });
+        }catch (IOException e){
             e.printStackTrace();
         }
         return dic;
     }
     public static void lsCommand() {
-        List<String> dic = listDirectory();
-        for(int i =1; i< dic.size();i++) {
-            System.out.println(dic.get(i));
+        Path path = Paths.get(Os_res.system_path+pwdCommand());
+        try(Stream<Path> list=Files.list(path)){
+        list.map(p -> {
+            if(Files.isDirectory(p)){
+                return "[D] " + p.getFileName();
+            }
+            else{
+                return "[F] " + p.getFileName();
+            }
+        }).forEach(System.out::println);
+
+        }catch (IOException e){
+            e.printStackTrace();
         }
+
     }
     public static void clearCommand() {
         for (int i = 0; i < 150; i++) {
@@ -95,12 +108,19 @@ public class Command_List {
 
     public static void rmCommand(String fileName) {
         Path path = Paths.get(Os_res.system_path + pwdCommand() + fileName);
+
         try {
+            if (!Objects.equals(fileName, "root")) {
             boolean isExist = Files.deleteIfExists(path);
-            if (!isExist)
-                System.out.println("file does not exist");
-            else
-                System.out.println("file has been deleted");
+                if (!isExist)
+                    System.out.println("file does not exist");
+                else
+                    System.out.println("file has been deleted");
+                }
+            else{
+                System.out.println("root can't be deleted");
+            }
+
         }
         catch (IOException e) {
             System.out.println("can't delete this file");
@@ -109,7 +129,8 @@ public class Command_List {
 
     public static void mvfCommand(String fileName) {
         List<String> directories = listDirectory();
-        for(int i=1;i<directories.size();i++) {
+
+        for(int i=0;i<directories.size();i++) {
             if(directories.get(i).equals(fileName)){
                 Os_res.os_path = Os_res.os_path + directories.get(i) + "/";
                 return;
@@ -119,12 +140,18 @@ public class Command_List {
     }
     public static void mvbCommand() {
         String[] current_paths = Os_res.os_path.split("/");
-        StringBuilder updated_paths = new StringBuilder();
-        for(int i=0;i<current_paths.length-1;i++) {
-            updated_paths.append(current_paths[i]).append("/");
-        }
-        Os_res.os_path = updated_paths.toString();
+        if(current_paths.length > 1 ){
+            StringBuilder updated_paths = new StringBuilder();
 
+            for (int i = 0; i < current_paths.length - 1; i++) {
+                updated_paths.append(current_paths[i]).append("/");
+            }
+
+            Os_res.os_path = updated_paths.toString();
+        }
+        else {
+            System.out.println("no path exists");
+        }
     }
